Fail if MTU cannot be read correctly
Summary:
Fail early if the device MTU is not readable during startup to prevent mysterious
pkt drops later on.
The PGW takes a device name to bind to for the SGi interface (the interface to the
outside world). In order to generate the right packets, the MTU of the device is
discovered by probing. In situations where the device is misconfigured the MTU
might not be discoverable. Check for such conditions and terminate instead of
mysteriously dropping packets due to mismatched MTUs at later points.

Test Plan:
Verified that we die early with a suitable error message in a misconfigured
setup.
Verified that we come up fine when a veth device is configured correctly.

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>