Magic +1 to snprintf length (for null terminator)
Summary:
Added +1 to n-value of snprintf so that entire IMSI value can get
written. Previously, a too-small n-value caused the IMSI value's last
digit to get cut off because snprintf needs to include the null
terminator in the length.

Test Plan: Run MME and see if the IMSI value is correct now (length and value).

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>