Fix ipv6 related buffer overflow issue
Summary:
The underlying socket for sctp is created using AF_INET6, which results in
a buffer override if the address is of type sockaddr_in vs sockadd_in6.

since fromlen is a value-result parameter, a reasonable behavior would
have been to do something sane if fromlen < size of src address but that
doesn't seem to be the case.
https://linux.die.net/man/3/sctp_recvmsg

Test Plan: Running some integration tests, seems fine now.

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>