Prevent stack overwrite while accessing IMSI (empty commit)
Summary:
emm_ue_id is a uint32_t variable allocated on the stack.
&emm_ue_id is a uintptr_t tmp containing the stack address for emm_ue_id,
This is typecast to void** which interprets the address pointed to by the tmp as a
pointer and hence of type uintptr_t which is > sizeof emm_ue_id (typedef to uint32_t).
This may result in a stack overwrite.
The hashtable interface actually expects a pointer to the data being read, so update
the code to do the same.

Test Plan: Running tests that seem to be going fine.

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>