Fix insertion pattern for IMSI
Summary:
Pass a pointer into the hash table as the hash table expects to manage
the data associated with an IMSI.

Seems to be working for now, the whole memory management around emm
seems to be reasonably complicated, so it is not obvious if something is
going to break, but we are not making the memory model any worse.

A better model would be to have the hash table "clone" the data and own it
for the duration of the hash key's validity. That way callers need to worry about
their memory lifecycle and the hash table manages its own memory.

Test Plan: s1aptester

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>