Handling the UE Context Release Request with Cause User Inactivity
Summary:
As of now , in MME UE Context Release Request with Cause User Inactivity  is ignored . This is not a correct behavior . MME should handle it.

Fix - Remove the check and remove the ignoring portion of logic in the code.

Note - W/o this fix , MME will not respond with UE Context Release Command .

Test Plan:
1. Attach the UE
2. Let inactivity timer expire in S1APteste or eNB( baicell or amarisoft ) .
3. This would trigger S1AP: UE Context release Request message from S1APTester or eNB with cause : Radio Network: User inactivity

Expected result - MME should not ignore this and should respond with UE Context Release Command.

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>