Fix EMM integrity protection
Summary:
Currently integrity protection doesn't work beyond the first message.
This causes subsequent NAS messages like DetacthAccept to be
dropped.
Fix the logic for computing the integrity to use bitwise or as opposed
to a logical or.
Overflow is not handled will fix that in a new change
as that is more of a feature add than a bug fix.

Test Plan: Verified detach works.

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>