Fix sctp handling in MME preventing double frees
Summary:
SCTP events are not handled correctly in the MME
  - Only enable events that we care about: association changes and shutdown.
  - Handle all association changes, currently only new associations are handled
    This results in a double free when an association closes and a new association
    with the same id gets recreated. See task for trace.

Cleanup doesn't do anything useful yet (the code is commented out), will fix that
in a subsequent change.

Test Plan:
Lots of weirdness happens when we restart an sctp association or just plain crash
the eNB and restart it. These are gone after this change.

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>