Caching radio capabilities data in MME
Summary:
ENBs currently locally cache UE radio capabilities data, but this gets
wiped from ENB each time the UE disconnects, requiring multiple requests
to the UE to retrieve this data for frequent disruptions. This can be
very bandwidth-consuming.

This caching allows MME to cache the radio capabilities data from ENBs,
so that upon reconnection to a UE, ENBs can retrieve the capabilities
from MME rather than the UE, thus saving on bandwidth use.

This change also implements the reset that can be bundled in the
TrackingAreaUpdateRequest from the UE to the MME, as well as a reset
built into the attach accept routine.

Test Plan:
Manual testing: test on physical devices. Test using S1AP tester once it
supports the communication features necessary to disconnect the UE to
idle to test the connected -> idle -> connected pathway.

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>