Cleanup on non-switchoff UE-initiated detach
Summary:
Cleanup on non-switchoff of UE initiated detach, today it is done
for only the power off case.
Also,
- Change the usage of 32 bit ue_ids typecast as 64 bit addresses.
- Update the freekeyfunc signature (regression introduced when I
modified the free function.

Test Plan:
Integration tests with the same UE.
Since we never hit the seg fault with the obj_hashtable_ts_remove after the
change with free_wrapper we are probably leaking memory for all the obj_hash tables.

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>