MME seg fault during 32-UE reconnect
Summary:
********Fix for MME seg fault during 32-UE Reconnect ********

As per the syslog/backtrace shared in the task#15954571, assert was happening during Initial UE message handling when UE reconnects with GUTI. This implied that while releasing S1 context for the UE , UE state was not moved to idle and enb_ue_s1ap_id was not made invalid .

Fixed few cases where it could happen. We dont have enough logs to be sure that issue is completely fixed. Need to retest to be sure .

Test Plan:
retest - same scenario as reported in task #15954571.

Verified multi UE attach-detach . Attached pcap log.
{F65717750}

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>