Fix a MME crash while resending attach_accept
Summary:
The TAI list was not getting copied over as part of the attach accept resend resulting in garbage
memory access. Copy over the datastructure even for the resend case to prevent the attach.
Also, fix a minor crash for the corner case where an attach request comes in before the attach
complete is recvd. by the UE.

Test Plan: see attached test

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>