Possible fix for heap overflow crash
Summary:
The logs are truncated so this is a speculative fix.
A possible reason for why we write past the buffer is that the
default_apn_p->nb_ip_address is non zero for the ue_context
when we are sending the s11_create_session_req but there are
no associated IP addresses
This could be due to the fact that apn parsing in s6a doesn't initialize
the nb_ip_address part of the struct.
Fix this by initializing the struct.
Even if this is not the fix that solves the overflow this is a safe and correct
fix.

Test Plan: s1aptests

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>