Add a test for sync and pass rand to subscriberd
Change is modified to accomodate the fact that the HSS doesn't expect the
RAND in the message.

Summary:
Change does three things:
a - Add support for RAND to be non zero by having the MME pass the stored rand value
b - Simplify the code in AttachComplete that solves for cases that shouldn't happen

Also added a unit test for the same.

Test Plan: test_resync and test_attach_detach

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>