IPv4 alloc defaults to NAS signaling in the absence of UE preference
Summary:
Allocating IPv4 addresses in SGW now uses NAS signaling to allocate an
IPv4 address if the UE did not specify a preference.

Created data structure to better carry the preferences from PGW to SGW
for purposes of checking UE preferences, in preparation for a more
complete solution.

Test Plan:
Run SGW and ensure that IPv4 is allocated even in absence of UE pref,
e.g. on TM500.

Signed-off-by: Anoop Tomar <anoop.tomar@radisys.com>