Update PID file location
Previously the mme pid file was always saved at location /var/run. In
situations where the MME is not running as a privileged user, this location
is not writable, so a config param was introduced for the PID file location.
Updated the checked in sample mme.conf to factor in this change.